get rid of int type and use most buffer with uint32_t

+++ openair1/PHY/CODING/nrLDPC_decoder/nrLDPC_CUDA_BnProcKernel_BG1.h
@@ -15,10 +15,10 @@ __device__ __forceinline__ void bnProcKernelMerge_BG1_int8_NUM(
     int8_t *__restrict__ d_cnProcBuf,
     const int8_t *__restrict__ d_llrProcBuf,
     int8_t *__restrict__ d_llrRes,
-    int lane,
-    int MsgIdx,
-    int BnIdx,
-    int GrpNum,
+    uint32_t lane,
+    uint8_t MsgIdx,
+    uint32_t BnIdx,
+    uint8_t GrpNum,
     uint32_t circShift,
     uint32_t Zc)
 {
@@ -29,12 +29,12 @@ __device__ __forceinline__ void bnProcKernelMerge_BG1_int8_NUM(
 //    int8_t *p_llrRes_BnIdx           = d_llrRes + baseBn;
 
     const int32_t *bnProcBufPtr = (const int32_t *)(d_bnProcBuf) + lane;
-    int prevIdxWords = ((MsgIdx - 1) * GrpNum * Zc) >> 2;
+    int32_t prevIdxWords = ((MsgIdx - 1) * GrpNum * Zc) >> 2;
     int32_t prev = bnProcBufPtr[prevIdxWords];
 
     // ---- ① Unrolled accumulation ----
     int32_t MsgSum = bnProcBufPtr[0];
-    int off = (GrpNum * Zc) >> 2;
+    int32_t off = (GrpNum * Zc) >> 2;
 #pragma unroll
     for (int i = 1; i < NUM; ++i) {
 	bnProcBufPtr += off;
@@ -59,11 +59,11 @@ __device__ __forceinline__ void bnProcKernelMerge_BG1_int8_Gn(
     int8_t *__restrict__ d_cnProcBuf,
     const int8_t *__restrict__ d_llrProcBuf,
     int8_t *__restrict__ d_llrRes,
-    int lane,
-    int GrpIdx,
-    int MsgIdx,
-    int BnIdx,
-    int GrpNum,
+    uint32_t lane,
+    uint8_t GrpIdx,
+    uint8_t MsgIdx,
+    uint32_t BnIdx,
+    uint8_t GrpNum,
     uint32_t circShift,
     uint32_t Zc)
 {
